extracted patches for custom spell support to separate class
